+++ src/models/download-data.ts
@@ -2,6 +2,7 @@ interface DownloadData {
   url: string;
   referer: string | null;
   pageAddress: string | null;
+  isBrowserNative: boolean;
 }
 
 export type { DownloadData };